Move TranspileReturn struct to transpile.rs
